[`refined4s-circe`] Update test structure for generic.auto - separate custom type test code from autoSpec - CustomTypeSpec.scala
